requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

torch==2.5.1
torchvision==0.20.1
timm==0.9.12

opencv-python-headless==4.8.1.78
Pillow==10.1.0
albumentations==1.3.1

numpy==1.26.0
scikit-learn==1.3.2
scikit-image==0.22.0
scipy==1.11.4